Removed unnecessary React dependencies
