+++ packages/react-querybuilder/src/redux/selectors.ts
@@ -1,4 +1,4 @@
-import type { RuleGroupTypeAny } from '../types';
+import type { RuleGroupTypeAny } from '../types/index.noReact';
 import { queriesSlice } from './queriesSlice';
 import type { RqbState } from './types';
